mini-projects/04-documentation-agent/pyproject.toml
[project]
name = "04-documentation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitsource>=0.0.4",
    "jaxn>=0.0.5",
    "minsearch>=0.0.11",
    "openai>=2.36.0",
    "pydantic-ai>=1.92.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.57.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
